Allow a single retry with wrong msg_type
Getters may fail as the VehicleService sends
messages periodically, and the expected message
with the value may not be the first one to be received.

To fix it, allow one retry. One should be enough
in most cases. The downside of this is that the
unexpected message is discarded. Therefore,
getters shall be used with care.

Alternatively, we could have getters that
discharge the responsibility of processing
the response to users.

Signed-off-by: Albert Esteve <aesteve@redhat.com>